# pos/openapi/RmsSupportHub.Pos.Agent.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RMS+ POS Agent API",
    "description": "Destination-owned foundation contract for the per-device Windows POS Agent. The Agent exposes a fixed HTTPS loopback endpoint for direct Support Hub browser access. Windows Negotiate supplies the authenticated identity, local Built-in Administrators membership authorizes protected reads and typed Windows-service and RMS database operations, and no Support Hub backend relay is involved. The Agent exposes read-only device/configuration/RMS visibility plus allow-listed service control and server-owned RMS database backup/restore workflows. Downloader, maintenance, and artifact capabilities remain typed, principal-scoped, and fail-closed. Mutation tokens are short-lived, one-use, and bound to the authenticated principal, exact Origin, registered operation, target, method, and path.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://rms-pos-agent.localhost:5001",
      "description": "Fixed HTTPS loopback Agent origin"
    }
  ],
  "paths": {
...
    "/api/v1/repair/preview/{operationId}": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Preview a typed Repair Installation operation",
        "description": "Validates the server-owned package manifest, signature, checksum, compatibility, archive, fixed installation boundary, required Safety Snapshot, capacity, and rollback preconditions without changing the machine.",
        "operationId": "PreviewRepairWithoutSnapshot",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the typed RepairPreviewDto precondition result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairPreviewDto"
                },
                "example": {
                  "previewId": "opaque-preview-id",
                  "operation": "repair",
                  "ready": false,
                  "packageVerification": "unknown",
                  "blockers": [
                    "package_unavailable",
                    "safety_snapshot_required"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/preview/{operationId}/{snapshotId}": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Preview a typed Repair Installation operation",
        "description": "Validates the server-owned package manifest, signature, checksum, compatibility, archive, fixed installation boundary, required Safety Snapshot, capacity, and rollback preconditions without changing the machine.",
        "operationId": "PreviewRepairWithSnapshot",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the typed RepairPreviewDto precondition result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairPreviewDto"
                },
                "example": {
                  "previewId": "opaque-preview-id",
                  "operation": "repair",
                  "ready": false,
                  "packageVerification": "unknown",
                  "blockers": [
                    "package_unavailable",
                    "safety_snapshot_required"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/operations": {
      "post": {
        "tags": [
          "Repair"
        ],
        "summary": "Start an authorized typed repair operation",
        "description": "Accepts only a bounded typed request after server-owned validation. Destructive or side-effecting dispatch requires the exact-origin, administrator, one-use method/path mutation-token, challenge/confirmation, and idempotency boundaries; ambiguous outcomes are retained and never retried automatically.",
        "operationId": "StartRepairOperation",
        "requestBody": {
          "description": "The browser supplies only logical IDs, exact confirmation where required, and a bounded idempotency key. Server-owned paths, credentials, SQL, and service targets are not accepted.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairExecuteRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The Agent accepted or safely rejected the typed repair request and reports activation, health, rollback, and recovery truth.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairOperationDto"
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected the typed request, challenge, configuration, or bounded idempotency key with safe problem details or typed NotAttempted state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge for this protected mutation.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, exact-origin transport, SID resolution, or the one-use mutation-token boundary rejected the mutation request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "The Agent returned a safe generic server-error response without exception, credential, path, or SQL details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/operations/{operationId}": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Read one principal-scoped repair operation",
        "description": "Reads principal-scoped, sanitized progress and outcome for one retained Agent operation. Operation IDs and artifact capabilities are opaque; raw paths, credentials, SQL, service names, and exception text never cross the response boundary.",
        "operationId": "GetRepairOperation",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the retained typed repair lifecycle state and recovery evidence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairOperationDto"
                },
                "example": {
                  "operationId": "opaque-operation-id",
                  "state": "running",
                  "progressPercent": 50
                }
              }
            }
          },
          "404": {
            "description": "The opaque operation identifier is not retained for the authenticated principal."
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/guided/preview": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Preview Guided Repair checkpoints",
        "description": "Returns the fixed typed checkpoint sequence and the next required precondition. A recommendation never implies package staging, activation, or a machine mutation.",
        "operationId": "PreviewGuidedRepair",
        "responses": {
          "200": {
            "description": "The Agent returned the typed GuidedRepairDto checkpoint projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidedRepairDto"
                },
                "example": {
                  "guidedRepairId": "opaque-guided-repair-id",
                  "state": "preview",
                  "steps": [ ],
                  "detail": "A fresh verified snapshot is required before guided repair."
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/guided/preview/{snapshotId}": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Preview Guided Repair checkpoints",
        "description": "Returns the fixed typed checkpoint sequence and the next required precondition. A recommendation never implies package staging, activation, or a machine mutation.",
        "operationId": "PreviewGuidedRepairWithSnapshot",
        "parameters": [
          {
            "name": "snapshotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the typed GuidedRepairDto checkpoint projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidedRepairDto"
                },
                "example": {
                  "guidedRepairId": "opaque-guided-repair-id",
                  "state": "preview",
                  "steps": [ ],
                  "detail": "A fresh verified snapshot is required before guided repair."
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repair/guided/{guidedRepairId}": {
      "get": {
        "tags": [
          "Repair"
        ],
        "summary": "Read one principal-scoped Guided Repair sequence",
        "description": "Reads principal-scoped, sanitized progress and outcome for one retained Agent operation. Operation IDs and artifact capabilities are opaque; raw paths, credentials, SQL, service names, and exception text never cross the response boundary.",
        "operationId": "GetGuidedRepair",
        "parameters": [
          {
            "name": "guidedRepairId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the retained typed GuidedRepairDto checkpoint state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidedRepairDto"
                },
                "example": {
                  "operationId": "opaque-operation-id",
                  "state": "running",
                  "progressPercent": 50
                }
              }
            }
          },
          "404": {
            "description": "The opaque operation identifier is not retained for the authenticated principal."
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RepairExecuteRequestDto": {
        "required": [
          "previewId",
          "typedConfirmation",
          "idempotencyKey"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "typedConfirmation": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Bounded caller-generated key used to prevent duplicate execution of the same typed request."
          },
          "snapshotId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          }
        },
        "description": "Typed request for one prior repair preview. It contains no installer path, command, or Main Server mutation input."
      },
...
      "RepairPreviewDto": {
        "required": [
          "previewId",
          "operation",
          "ready",
          "packageVerification",
          "snapshot",
          "effects",
          "blockers",
          "confirmationPhrase",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "operation": {
            "description": "Typed repair operation selected by the server-owned policy.",
            "$ref": "#/components/schemas/RepairOperationKindDto"
          },
          "ready": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "packageVerification": {
            "description": "Typed package verification state required before repair can proceed.",
            "$ref": "#/components/schemas/AgentPackageVerificationStateDto"
          },
          "snapshot": {
            "description": "Reference to the fresh principal-scoped Safety Snapshot required by this repair preview.",
            "$ref": "#/components/schemas/SafetySnapshotReferenceDto"
          },
          "effects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "confirmationPhrase": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "expiresAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          }
        },
        "description": "Typed repair precondition preview bound to a fresh principal-scoped Safety Snapshot."
      },
...
    },
    "securitySchemes": {
      "WindowsNegotiate": {
        "type": "http",
        "description": "Windows Integrated Authentication using Negotiate. The browser/OS supplies credentials with a credentialed cross-origin request; clients must rely on the browser's integrated credential flow.",
        "scheme": "negotiate"
      }
    }
  },
...
}